workspace/mediscope/apps/worker/requirements.txt
beautifulsoup4==4.14.3
jinja2==3.1.6
fastapi==0.135.2
httpx==0.28.1
lxml==6.0.2
playwright==1.58.0
pydantic-settings==2.13.1
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
uvicorn==0.42.0
